Intent: Clarify trig caching boundaries Operators involved: [Coherence, Coupling] Affected invariants: [#5, #8] Key changes: - Rename tnfr.metrics.trigonometry module to trig_cache and update imports - Document split between trig.py helpers and trig_cache cosine/sine storage - Align phase-related tests with the new module name Expected risks/dissonances: Minimal; import paths updated consistently and covered by tests Metrics: Phase means and Si computations remain stable (validated via targeted pytest suite) Equivalence map: tnfr.metrics.trigonometry -> tnfr.metrics.trig_cache
