+++ benchmarks/README.md
@@ -22,7 +22,7 @@ PYTHONPATH=src python benchmarks/<script>.py
 | --- | --- | --- |
 | `cached_abs_max.py` | Cache-aware updates for absolute maxima (`tnfr.alias.set_attr_with_max`). | Demonstrates how cached maxima avoid recomputing `recompute_abs_max` on every assignment. |
 | `collect_attr.py` | Vectorised collection of nodal attributes (`tnfr.alias.collect_attr`). | Requires NumPy; the script exits gracefully when the module is unavailable. |
-| `neighbor_phase_mean.py` | Fast phase averaging for neighbourhoods (`tnfr.metrics.trig.neighbor_phase_mean`). | Includes a `NodoNX`-based reference to highlight the benefit of cached trigonometry. |
+| `neighbor_phase_mean.py` | Fast phase averaging for neighbourhoods (`tnfr.metrics.trig.neighbor_phase_mean`). | Includes a `NodoNX`-based reference to highlight the benefit of the shared `trig_cache` module. |
 | `prepare_dnfr_data.py` | ΔNFR data preparation reuse (`tnfr.dynamics._prepare_dnfr_data`). | Exercises cache reuse when assembling phase/EPI/νf arrays. |
 
 ## Retired scripts
