+++ src/tnfr/dynamics/__init__.py
@@ -36,7 +36,7 @@ from ..alias import (
 )
 from ..metrics.sense_index import compute_Si
 from ..metrics.common import compute_dnfr_accel_max, merge_and_normalize_weights
-from ..metrics.trigonometry import compute_theta_trig
+from ..metrics.trig_cache import compute_theta_trig
 from ..callback_utils import callback_manager
 from ..glyph_history import recent_glyph, ensure_history, append_metric
 from ..selector import (
