+++ src/tnfr/metrics/coherence.py
@@ -18,7 +18,7 @@ from ..collections_utils import normalize_weights
 from ..helpers.numeric import clamp01
 from ..helpers import ensure_node_index_map
 from .common import compute_coherence, min_max_range
-from .trigonometry import compute_theta_trig, get_trig_cache
+from .trig_cache import compute_theta_trig, get_trig_cache
 from ..observers import glyph_load, kuramoto_order, phase_sync
 from ..sense import sigma_vector
 from ..import_utils import get_numpy
